Replace all_definitions_ map with vector
We know the id bound, so we can just allocate this upfront.

Saves about 15s on my profile.

TODO: How much do we care about poorly-written modules with very sparse
id usage? Some spirv consumers fall back to a hashed slow-path for that;
we could do that, or we could rewrite the ids to a dense range on the
fly, or we could declare it a non-problem.